+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 - [IMPROVED] Updated `Getting started` section with a `get_query_result` example. 
 - [FIXED] Fixed parameter type of `selector` in docstring.
+- [NEW] `Document.atomic` context manager.
 
 # 2.11.0 (2019-01-21)